Fix final Android compilation errors: Type conversion and scope issues
- JobsRepository: Convert orderId String to Int for UIDScanEntity constructor
- JobDetailScreen: Fix variable scope by using onStatusUpdate callback instead of undefined variables
- Complete order button now properly integrates with existing status update workflow

All Android compilation errors resolved for production build.

+++ driver-app/app/src/main/java/com/yourco/driverAA/domain/JobsRepository.kt
@@ -144,7 +144,7 @@ class JobsRepository @Inject constructor(
             // Store UID actions locally (for offline support)
             uidActions.forEach { uidAction ->
                 val uidScan = UIDScanEntity(
-                    orderId = orderId,
+                    orderId = orderId.toInt(),
                     uid = uidAction.uid,
                     action = uidAction.action,
                     skuId = uidAction.sku_id,
